% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjPlotAnova.R
\name{levene_test}
\alias{levene_test}
\title{Plot Levene-Test for One-Way-Anova}
\usage{
levene_test(depVar, grpVar)
}
\arguments{
\item{depVar}{The dependent variable. Will be used with following formular:
\code{aov(depVar ~ grpVar)}}

\item{grpVar}{The grouping variable, as unordered factor. Will be used with following formular:
         \code{aov(depVar ~ grpVar)}}
}
\description{
Plot results of Levene's Test for Equality of Variances for One-Way-Anova.
}
\examples{
data(efc)
levene_test(efc$c12hour, efc$e42dep)
}
\seealso{
\code{\link{sjp.aov1}}, \code{\link{chisq_gof}}, \code{\link{mwu}} and \code{\link{wilcox.test}},
         \code{\link{ks.test}}, \code{\link{kruskal.test}}, \code{\link{t.test}}, \code{\link{chisq.test}},
         \code{\link{fisher.test}}
}

