% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjStatistics.R
\name{phi}
\alias{phi}
\title{Phi value for a contingency table}
\usage{
phi(tab)
}
\arguments{
\item{tab}{A simple \code{\link{table}} or \code{\link{ftable}}. Tables of class
\code{\link{xtabs}} and other will be coerced to \code{\link{ftable}} objects.}
}
\value{
The table's Phi value.
}
\description{
Compute Phi value for a contingency table.
}
\examples{
tab <- table(sample(1:2, 30, TRUE), sample(1:2, 30, TRUE))
phi(tab)
}
\seealso{
\code{\link{table_values}} \cr
         \code{\link{cramer}}
}

