% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{read_sas}
\alias{read_sas}
\title{Import SAS dataset as data frame into R}
\usage{
read_sas(path, path.cat = NULL)
}
\arguments{
\item{path}{The file path to the SAS data file.}

\item{path.cat}{optional, the file path to the SAS catalog file.}
}
\value{
A data frame containing the SAS data. Retrieve value labels with \code{\link{get_val_labels}}
  and variable labels with \code{\link{get_var_labels}}.
}
\description{
Imports data from SAS (\code{.sas7bdat}), including NA's,
               value and variable labels.
}
\note{
This is a wrapper function for \code{read_sas} function of the
        \code{haven} package. This function converts the imported data
        into a sjPlot friendly format (see \code{\link{to_sjPlot}}).
}
\seealso{
\itemize{
           \item \code{\link{read_spss}}
           }
}

