% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjPlotSetTheme.R
\name{sj.setGeomColors}
\alias{sj.setGeomColors}
\title{Helper function to set geom colors}
\usage{
sj.setGeomColors(plot, geom.colors, pal.len, show.guide = TRUE,
  labels = NULL, reverse.colors = FALSE)
}
\arguments{
\item{plot}{a ggplot object where scales (geom colors) should be set}

\item{geom.colors}{the color palette for the scales to be used}

\item{pal.len}{the length of the required colors in \code{geom.colors}}

\item{show.guide}{whether or not legend should be displayed}

\item{labels}{a character vector with legend labels}

\item{reverse.colors}{If \code{TRUE}, the color scale is reversed.}
}
\description{
Helper function to set geom colors
}

