% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{sjc.qclus}
\alias{sjc.qclus}
\alias{sjc.cluster}
\alias{sjc.grpdisc}
\alias{sjc.elbow}
\alias{sjc.kgap}
\title{Deprecated functions}
\usage{
sjc.qclus(x, ...)

sjc.cluster(x, ...)

sjc.grpdisc(x, ...)

sjc.elbow(x, ...)

sjc.kgap(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Currently not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
