% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{sji.getValueLabels}
\alias{get_val_lab}
\alias{sji.getValueLabels}
\title{Retrieve value labels of a variable or an SPSS-imported data frame}
\usage{
sji.getValueLabels(x)

get_val_lab(x)
}
\arguments{
\item{x}{a data frame with variables that have attached value labels (e.g.
from an imported SPSS data (see \code{\link{sji.SPSS}})) or a variable
with attached value labels.}
}
\value{
Either a list with all value labels from the data frame's variables,
          or a string with the value labels, if \code{x} is a variable.
}
\description{
This function retrieves the value labels of an imported
               SPSS data set and returns the result as list or of a variable and returns
               the label as string.
}
\section{Functions}{
\itemize{
\item \code{get_val_lab}: 
}}
\note{
This function only works with vectors that have value and variable
       labels attached. This is automatically done by importing SPSS data sets
       with the \code{\link{read_spss}} function and can manually be applied
       with the \code{\link{sji.setValueLabels}} and \code{\link{sji.setVariableLabels}}
       functions.
}
\examples{
# import SPSS data set
# mydat <- sji.SPSS("my_spss_data.sav", enc="UTF-8")

# retrieve variable labels
# mydat.var <- sji.getVariableLabels(mydat)

# retrieve value labels
# mydat.val <- sji.getValueLabels(mydat)

data(efc)
sji.getValueLabels(efc$e42dep)
}
\seealso{
\itemize{
           \item \href{http://www.strengejacke.de/sjPlot/datainit/}{sjPlot manual: data initialization}
           \item \href{http://www.strengejacke.de/sjPlot/sji.viewSPSS/}{sjPlot manual: inspecting (SPSS imported) data frames}
           \item \code{\link{sji.SPSS}}
           \item \code{\link{sji.getVariableLabels}}
           \item \code{\link{sji.convertToLabel}}
           \item \code{\link{sji.convertToValue}}
           \item \code{\link{sji.setValueLabels}}
           }
}

