% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{sji.toFac}
\alias{sji.toFac}
\alias{to_fac}
\title{Convert variable into factor and keep value labels}
\usage{
sji.toFac(var)

to_fac(var)
}
\arguments{
\item{var}{A (numeric or atomic) variable.}
}
\value{
A factor variable, including variable and value labels.
}
\description{
This function converts a variable into a factor, but keeps
               variable and value labels, if these are attached as attributes
               to the variale \code{var}. See examples.
}
\section{Functions}{
\itemize{
\item \code{to_fac}: 
}}
\examples{
\dontrun{
data(efc)
# normal factor conversion, loses value attributes
efc$e42dep <- as.factor(efc$e42dep)
sjt.frq(efc$e42dep)

data(efc)
# factor conversion, which keeps value attributes
efc$e42dep <- sji.toFac(efc$e42dep)
sjt.frq(efc$e42dep)}
}
\seealso{
\itemize{
           \item \code{\link{sji.convertToValue}}
           \item \code{\link{sji.convertToLabel}}
           \item \code{\link{sji.getValueLabels}}
           \item \code{\link{sji.getVariableLabels}}
           \item \code{\link{sji.SPSS}}
           }
}

