% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{sjp.glmer}
\alias{sjp.glmer}
\alias{sjp.glm}
\alias{sjp.lmer}
\alias{sjp.lm}
\alias{sjp.int}
\alias{sjt.frq}
\alias{sjp.scatter}
\title{Deprecated functions}
\usage{
sjp.glmer(...)

sjp.glm(...)

sjp.lmer(...)

sjp.lm(...)

sjp.int(...)

sjt.frq(...)

sjp.scatter(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
