% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sjPlotGroupPropTable.R
\name{sjp.gpt}
\alias{sjp.gpt}
\title{Plot grouped proportional tables}
\usage{
sjp.gpt(x, y, groups, legendTitle, legendLabels)
}
\value{
(Insisibily) returns the ggplot-object with the complete plot (\code{plot}) as well as the data frame that
          was used for setting up the ggplot-object (\code{mydf}).
}
\description{
Plot proportional crosstables (contingency tables) of two variables as ggplot diagram.
}

