% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjPlotGroupFrequencies.R
\name{sjp.grpfrq}
\alias{sjp.grpfrq}
\title{Plot grouped or stacked frequencies}
\usage{
sjp.grpfrq(varCount, varGroup, weight.by = NULL, weightByTitleString = NULL,
  interactionVar = NULL, type = c("bar", "dot", "line", "boxplot",
  "violin"), geom.size = NULL, geom.spacing = 0.15,
  geom.colors = "Paired", show.legend = TRUE, facet.grid = FALSE,
  title = "", legendTitle = NULL, axis.labels = NULL,
  interactionVarLabels = NULL, legendLabels = NULL, ylim = NULL,
  breakTitleAt = 50, breakLabelsAt = 15, breakLegendTitleAt = 20,
  breakLegendLabelsAt = 20, grid.breaks = NULL, barPosition = "dodge",
  innerBoxPlotWidth = 0.15, innerBoxPlotDotSize = 3, smoothLines = FALSE,
  expand.grid = FALSE, show.values = TRUE, show.n = TRUE,
  show.perc = TRUE, show.axis.values = TRUE, showPlotAnnotation = TRUE,
  showTableSummary = FALSE, showGroupCount = FALSE, tableSummaryPos = "r",
  axis.titles = NULL, autoGroupAt = NULL, coord.flip = FALSE,
  vjust = "bottom", hjust = "center", y.offset = NULL, na.rm = TRUE,
  printPlot = TRUE)
}
\arguments{
\item{varCount}{a vector of values (variable) describing the bars which make up the plot.}

\item{varGroup}{grouping variable of same length as \code{varCount}, where \code{varCount} 
is grouped into the categories represented by \code{varGrp}.}

\item{weight.by}{weight factor that will be applied to weight all cases.
Must be a vector of same length as the input vector. Default is 
\code{NULL}, so no weights are used.}

\item{weightByTitleString}{suffix (as string) for the plot's title, if \code{weight.by} is specified,
e.g. \code{weightByTitleString=" (weighted)"}. Default is \code{NULL}, so plot's 
title will not have a suffix when cases are weighted.}

\item{interactionVar}{an interaction variable which can be used for box plots. Divides each category indicated
by \code{varGroup} into the factors of \code{interactionVar}, so that each category of \code{varGroup}
is subgrouped into \code{interactionVar}'s categories. Only applies when argument \code{type}
is \code{box} or \code{violin} (resp. their alternative strings like \code{"boxplot"}, \code{"boxplot"} or \code{"v"}).}

\item{type}{Specifies the plot type. May be abbreviated.
\describe{
  \item{\code{"bar"}}{for simple bars (default)}
  \item{\code{"dot"}}{for a dot plot}
  \item{\code{"histogram"}}{for a histogram (does not apply to grouped frequencies)}
  \item{\code{"line"}}{for a line-styled histogram with filled area}
  \item{\code{"density"}}{for a density plot (does not apply to grouped frequencies)}
  \item{\code{"boxplot"}}{for box plot}
  \item{\code{"violin"}}{for violin plots}
  }}

\item{geom.size}{size resp. width of the geoms (bar width, line thickness or point size, 
depending on plot type and function. Note that bar and bin widths mostly 
need smaller values than dot sizes.}

\item{geom.spacing}{the spacing between geoms (i.e. bar spacing)}

\item{geom.colors}{user defined color for bars. See 'Details' in \code{\link{sjp.grpfrq}}.}

\item{show.legend}{logical, if \code{TRUE}, and depending on plot type and
function, a legend is added to the plot.}

\item{facet.grid}{\code{TRUE} when bar charts should be plotted as facet grids instead of integrated single
bar charts. Ideal for larger amount of groups. This argument wraps a single panel into 
\code{varGroup} amount of panels, i.e. each group is represented within a new panel.}

\item{title}{string, used as plot title. Depending on plot type and function,
will be set automatically. If \code{title = ""}, no title is printed.}

\item{legendTitle}{title of the plot legend, as string.}

\item{axis.labels}{character vector with labels used as axis labels.}

\item{interactionVarLabels}{a character vector with labels for the x-axis breaks
when having interaction variables included.
These labels replace the \code{axis.labels}. Only applies, when using box or violin plots
(i.e. \code{type = "boxplot"} or \code{"violin"}) and \code{interactionVar} is not \code{NULL}.
Example: See \code{axis.labels}.}

\item{legendLabels}{a character vector with labels for the guide/legend.}

\item{ylim}{numeric vector of length two, defining lower and upper axis limits
of the y scale. By default, this argument is set to \code{NULL}, i.e. the 
y-axis fits to the required range of the data.}

\item{breakTitleAt}{determines how many chars of the plot title are displayed in
one line and when a line break is inserted into the title.}

\item{breakLabelsAt}{determines how many chars of the category labels are displayed in 
one line and when a line break is inserted.}

\item{breakLegendTitleAt}{determines how many chars of the legend's title 
are displayed in one line and when a line break is inserted.}

\item{breakLegendLabelsAt}{determines how many chars of the legend labels are 
displayed in one line and when a line break is inserted.}

\item{grid.breaks}{numeric; sets the distance between breaks for the axis, 
i.e. at every \code{grid.breaks}'th position a major grid is being printed.}

\item{barPosition}{indicates whether bars should be positioned side-by-side (default)
or stacked (use \code{"stack"} as argument).}

\item{innerBoxPlotWidth}{width of the inner box plot that is plotted inside of violin plots. Only applies 
if \code{type = "violin"}. Default value is 0.15}

\item{innerBoxPlotDotSize}{size of mean dot insie a violin or box plot. Applies only 
when \code{type = "violin"} or \code{"boxplot"}.}

\item{smoothLines}{prints a smooth line curve. Only applies, when argument \code{type = "line"}.}

\item{expand.grid}{logical, if \code{TRUE}, the plot grid is expanded, i.e. there is a small margin between
axes and plotting region. Default is \code{FALSE}.}

\item{show.values}{logical, whether values should be plotted or not.}

\item{show.n}{logical, if \code{TRUE}, adds total number of cases for each
group or category to the labels.}

\item{show.perc}{logical, if \code{TRUE} (default), percentage values are plotted to each bar
If \code{FALSE}, percentage values are removed.}

\item{show.axis.values}{logical, whether count or percentage values for the axis
should be printed or not.}

\item{showPlotAnnotation}{logical, if \code{TRUE}, the groups of dots in a dot-plot are highlighted 
with a shaded rectangle.}

\item{showTableSummary}{logical, if \code{TRUE}, a summary of the cross tabulation with N, 
chi-squared, df, Cramer's V or Phi-value and p-value is printed to the upper 
right corner of the plot (see \code{tableSummaryPos}. If a cell contains expected
values lower than five(or lower than 10 if df is 1), the Fisher's excact test 
(see \code{\link{fisher.test}}) is computed instead of chi-square test. 
If the table's matrix is larger than 2x2, Fisher's excact test with Monte Carlo 
simulation is computed. Only applies to barcharts or dotplots, i.e. 
when argument \code{type = "bar"} or \code{"dot"}.}

\item{showGroupCount}{logical, if \code{TRUE}, the count within each group is added 
to the category labels (e.g. \code{"Cat 1 (n=87)"}). Default value is \code{FALSE}.}

\item{tableSummaryPos}{position of the model summary which is printed when \code{showTableSummary} 
is \code{TRUE}. Default is \code{"r"}, i.e. it's printed to the upper right corner. 
Use \code{"l"} for upper left corner.}

\item{axis.titles}{character vector of length one or two, defining the title(s)
for the x-axis and y-axis.}

\item{autoGroupAt}{numeric value, indicating at which length of unique values of \code{varCount}, 
automatic grouping into smaller units is done (see \code{\link[sjmisc]{group_var}}).
If \code{varCount} has large numbers of unique values, there may be too many bars 
for the plot. Hence it's practical to group such variables. For example, 
if \code{autoGroupAt = 50} and \code{varCount} has more than 50 unique values,
it will be grouped (using the \code{\link[sjmisc]{group_var}} function). 
Default value for \code{autoGroupAt} is \code{NULL}, i.e. auto-grouping is off.
See \code{\link[sjmisc]{group_var}} for examples on grouping.}

\item{coord.flip}{logical, if \code{TRUE}, the x and y axis are swapped.}

\item{vjust}{character vector, indicating the vertical position of value 
labels. Allowed are same values as for \code{vjust} aesthetics from 
\code{ggplot2}: "left", "center", "right", "bottom", "middle", "top" and
new options like "inward" and "outward", which align text towards and 
away from the center of the plot respectively.}

\item{hjust}{character vector, indicating the horizontal position of value 
labels. Allowed are same values as for \code{vjust} aesthetics from 
\code{ggplot2}: "left", "center", "right", "bottom", "middle", "top" and
new options like "inward" and "outward", which align text towards and 
away from the center of the plot respectively.}

\item{y.offset}{numeric, offset for text labels when their alignment is adjusted 
to the top/bottom of the geom (see \code{hjust} and \code{vjust}).}

\item{na.rm}{logical, if \code{TRUE}, missings are not included in the frequency plot.}

\item{printPlot}{logical, if \code{TRUE} (default), plots the results as graph. Use \code{FALSE} if you don't
want to plot any graphs. In either case, the ggplot-object will be returned as value.}
}
\value{
(Insisibily) returns the ggplot-object with the complete plot (\code{plot}) as well as the data frame that
          was used for setting up the ggplot-object (\code{df}).
}
\description{
Plot grouped or stacked frequencies of variables as bar/dot, 
               box or violin plots, or line plot.
}
\details{
\code{geom.colors} may be a acharacter vector of color values 
         in hex-format, or a name of a \href{http://colorbrewer2.org}{color brewer} palette.
         Following options are valid for the \code{geom.colors} argument:
         \itemize{
           \item If not specified, a default color brewer palette will be used, which is suitable for the plot style (i.e. diverging for likert scales, qualitative for grouped bars etc.).
           \item If \code{"gs"}, a greyscale will be used.
           \item If \code{geom.colors} is any valid color brewer palette name, the related palette will be used. Use \code{\link[RColorBrewer]{display.brewer.all}} to view all available palette names.
           \item Else specify own color values as vector (e.g. \code{geom.colors = c("#f00000", "#00ff00")}).
         }
}
\examples{
# histrogram with EUROFAMCARE sample dataset
library(sjmisc)
data(efc)
sjp.grpfrq(efc$e17age,
           efc$e16sex,
           show.values = FALSE)

# boxplot
sjp.grpfrq(efc$e17age, 
           efc$e42dep, 
           type = "box")

# -------------------------------------------------
# auto-detection of value labels and variable names
# -------------------------------------------------
# grouped bars using necessary y-limit            
sjp.grpfrq(efc$e42dep, 
           efc$e16sex, 
           title = NULL)

# box plots with interaction variable            
sjp.grpfrq(efc$e17age,
           efc$e42dep,
           interactionVar = efc$e16sex,
           type = "box")

# Grouped bar plot ranging from 7 to 28
sjp.grpfrq(efc$neg_c_7, 
           efc$e42dep, 
           show.values = FALSE)

# same data as line plot
sjp.grpfrq(efc$neg_c_7, 
           efc$e42dep, 
           type = "line")
           
}
\seealso{
\href{http://www.strengejacke.de/sjPlot/sjp.grpfrq/}{sjPlot manual: sjp.grpfrq}
}

