% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjPlotResiduals.R
\name{sjp.resid}
\alias{sjp.resid}
\title{Plot predicted values and their residuals}
\usage{
sjp.resid(fit, remove.estimates = NULL, show.lines = TRUE,
  show.resid = TRUE, show.pred = TRUE, prnt.plot = TRUE)
}
\arguments{
\item{fit}{fitted linear regression model (of class \code{\link{lm}}, 
\code{\link[nlme]{gls}} or \code{plm}).}

\item{remove.estimates}{character vector with coefficient names that indicate
which estimates should be removed from the plot.
\code{remove.estimates = "est_name"} would remove the estimate \emph{est_name}. Default
is \code{NULL}, i.e. all estimates are printed.}

\item{show.lines}{logical, if \code{TRUE}, a line connecting predicted and
residual values is plotted. Set this argument to \code{FALSE}, if
plot-building is too time consuming.}

\item{show.resid}{logical, if \code{TRUE}, residual values are plotted.}

\item{show.pred}{logical, if \code{TRUE}, predicted values are plotted.}

\item{prnt.plot}{logical, if \code{TRUE} (default), plots the results as graph. Use \code{FALSE} if you don't
want to plot any graphs. In either case, the ggplot-object will be returned as value.}
}
\value{
(Insisibily) returns the ggplot-object with the complete plot (\code{plot}) 
          as well as the data frame that
          was used for setting up the ggplot-object (\code{mydf}).
}
\description{
to follow...
}
\examples{
library(sjmisc)
data(efc)
# fit model
fit <- lm(neg_c_7 ~ c12hour + e17age + e42dep, data = efc)

# plot residuals for all independent variables
sjp.resid(fit)

# remove some independent variables from output
sjp.resid(fit, remove.estimates = c("e17age", "e42dep"))

}

