% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sjs.cramer}
\alias{sjs.cramer}
\title{Cramer's V for a contingency table}
\usage{
sjs.cramer(tab)
}
\arguments{
\item{tab}{A simple \code{\link{table}} or \code{\link{ftable}}. Tables of class
\code{\link{xtabs}} and other will be coerced to \code{\link{ftable}} objects.}
}
\value{
The table's Cramer's V.
}
\description{
Compute Cramer's V for a table with more than 2x2 fields.
}
\examples{
tab <- table(sample(1:2, 30, TRUE), sample(1:3, 30, TRUE))
sjs.cramer(tab)
}
\seealso{
\code{\link{sjs.table.values}} \cr
         \code{\link{sjs.phi}}
}

