% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sjs.phi}
\alias{sjs.phi}
\title{Phi value for a contingency table}
\usage{
sjs.phi(tab)
}
\arguments{
\item{tab}{A simple \code{\link{table}} or \code{\link{ftable}}. Tables of class
\code{\link{xtabs}} and other will be coerced to \code{\link{ftable}} objects.}
}
\value{
The table's Phi value.
}
\description{
Compute Phi value for a contingency table.
}
\examples{
tab <- table(sample(1:2, 30, TRUE), sample(1:2, 30, TRUE))
sjs.phi(tab)
}
\seealso{
\code{\link{sjs.table.values}} \cr
         \code{\link{sjs.cramer}}
}

