% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabFA.R
\name{sjt.fa}
\alias{sjt.fa}
\title{Summary of factor analysis as HTML table}
\usage{
sjt.fa(data, rotation = c("promax", "varimax"), method = "ml",
  nmbr.fctr = NULL, fctr.load.tlrn = 0.1, title = "Factor Analysis",
  var.labels = NULL, wrap.labels = 40, show.cronb = TRUE,
  show.comm = FALSE, altr.row.col = FALSE, digits = 2, CSS = NULL,
  encoding = NULL, file = NULL, use.viewer = TRUE, no.output = FALSE,
  remove.spaces = TRUE)
}
\arguments{
\item{data}{\code{data.frame} that should be used to compute a FA, or a \code{\link[psych]{fa}} object.}

\item{rotation}{rotation of the factor loadings. May be \code{"varimax"} for orthogonal rotation
or \code{"promax"} for oblique transformation (default). Requires the \code{"GPArotation"} package.}

\item{method}{the factoring method to be used. "ml" will do a maximum likelihood factor analysis (default). 
"minres" will do a minimum residual (OLS), 
"wls" will do a weighted least squares (WLS) solution, 
"gls" does a generalized weighted least squares (GLS), 
"pa" will do the principal factor solution, 
"minchi" will minimize the sample size weighted chi square when treating pairwise 
correlations with different number of subjects per pair. 
minrank" will do a minimum rank factor analysis.}

\item{nmbr.fctr}{number of factors used for calculating the varimax
rotation. By default, this value is \code{NULL} and the amount of factors is
calculated according to a parallel analysis.}

\item{title}{table caption, as character vector.}

\item{var.labels}{character vector with variable names, which will be used 
to label variables in the output.}

\item{wrap.labels}{numeric, determines how many chars of the value, variable or axis 
labels are displayed in one line and when a line break is inserted.}

\item{show.cronb}{logical, if \code{TRUE} (default), the cronbach's alpha value for each factor scale will be calculated,
i.e. all variables with the highest loading for a factor are taken for the
reliability test. The result is an alpha value for each factor dimension.
Only applies when \code{data} is a data frame and no \code{\link[psych]{fa}} object.}

\item{show.comm}{logical, if \code{TRUE}, shows an additional column with item communalities.}

\item{altr.row.col}{logical, if \code{TRUE}, alternating rows are highlighted with a light gray
background color.}

\item{digits}{numeric, amount of digits after decimal point when rounding estimates and values.}

\item{CSS}{\code{\link{list}}-object with user-defined style-sheet-definitions, according to the 
\href{http://www.w3.org/Style/CSS/}{official CSS syntax}. For more details,
see \href{../doc/sjtbasic.html}{this package-vignette}, or 'Details' in
\code{\link{sjt.frq}}.}

\item{encoding}{string, indicating the charset encoding used for variable and 
value labels. Default is \code{NULL}, so encoding will be auto-detected 
depending on your platform (e.g., \code{"UTF-8"} for Unix and \code{"Windows-1252"} for
Windows OS). Change encoding if specific chars are not properly displayed (e.g. German umlauts).}

\item{file}{destination file, if the output should be saved as file.
If \code{NULL} (default), the output will be saved as temporary file and 
openend either in the IDE's viewer pane or the default web browser.}

\item{use.viewer}{If \code{TRUE}, the HTML table is shown in the IDE's viewer pane. If
\code{FALSE} or no viewer available, the HTML table is opened in a web browser.}

\item{no.output}{logical, if \code{TRUE}, the html-output is neither opened in a browser nor shown in
the viewer pane and not even saved to file. This option is useful when the html output
should be used in \code{knitr} documents. The html output can be accessed via the return
value.}

\item{remove.spaces}{logical, if \code{TRUE}, leading spaces are removed from all lines in the final string
that contains the html-data. Use this, if you want to remove parantheses for html-tags. The html-source
may look less pretty, but it may help when exporting html-tables to office tools.}
}
\value{
Invisibly returns
         \itemize{
           \item the web page style sheet (\code{page.style}),
           \item the web page content (\code{page.content}),
           \item the complete html-output (\code{output.complete}),
           \item the html-table with inline-css for use with knitr (\code{knitr}),
           \item the \code{factor.index}, i.e. the column index of each variable with the highest factor loading for each factor and
           \item the \code{removed.items}, i.e. which variables have been removed because they were outside of the \code{fctr.load.tlrn}'s range.
           }
           for further use.
}
\description{
Performes a factor analysis on a data frame or matrix 
               and displays the factor solution as HTML 
               table, or saves them as file. \cr In case a data frame is used as 
               parameter, the Cronbach's Alpha value for each factor scale will be calculated,
               i.e. all variables with the highest loading for a factor are taken for the
               reliability test. The result is an alpha value for each factor dimension.
}
\details{
See 'Details' in \code{\link{sjt.frq}}.
}
\note{
See 'Notes' in \code{\link{sjt.frq}}.
       This method for factor analysis relies on the functions  
\code{\link[psych]{fa}} and  \code{\link[psych]{fa.parallel}} from the psych package.
}
\examples{
\dontrun{
# Data from the EUROFAMCARE sample dataset
library(sjmisc)
data(efc)

# retrieve variable and value labels
varlabs <- get_label(efc)

# recveive first item of COPE-index scale
start <- which(colnames(efc) == "c82cop1")
# recveive last item of COPE-index scale
end <- which(colnames(efc) == "c90cop9")
 
# create data frame with COPE-index scale
mydf <- as.data.frame(efc[, c(start:end)])
colnames(mydf) <- varlabs[c(start:end)]

sjt.fa(mydf)

# auto-detection of labels
sjt.fa(efc[, c(start:end)])}

}
\seealso{
\itemize{
           \item \href{http://www.strengejacke.de/sjPlot/sjt.pca/}{sjPlot manual: sjt.pca}
           \item \code{\link{sjp.pca}}
         }
}
