% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabFlatTable.R
\name{sjt.ftab}
\alias{sjt.ftab}
\title{Flat (proportional) tables}
\usage{
sjt.ftab(.data, ..., show.obs = TRUE, show.cell.prc = FALSE,
  show.row.prc = FALSE, show.col.prc = FALSE, digits = 2,
  tdcol.n = "black", tdcol.expected = "#339999", tdcol.cell = "#993333",
  tdcol.row = "#333399", tdcol.col = "#339933", emph.total = FALSE,
  emph.color = "#f8f8f8", prc.sign = "&nbsp;&#37;", hundret = "100.0",
  CSS = NULL, encoding = NULL, file = NULL, use.viewer = TRUE,
  no.output = FALSE, remove.spaces = TRUE)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{One or more variables of \code{.data} that should be printed as table.}

\item{show.obs}{logical, if \code{TRUE}, observed values are shown}

\item{show.cell.prc}{logical, if \code{TRUE}, cell percentage values are shown}

\item{show.row.prc}{logical, if \code{TRUE}, row percentage values are shown}

\item{show.col.prc}{logical, if \code{TRUE}, column percentage values are shown}

\item{digits}{Numeric; for proportional tables, \code{digits} indicates the
number of decimal places.}

\item{tdcol.n}{Color for highlighting count (observed) values in table cells. Default is black.}

\item{tdcol.expected}{Color for highlighting expected values in table cells. Default is cyan.}

\item{tdcol.cell}{Color for highlighting cell percentage values in table cells. Default is red.}

\item{tdcol.row}{Color for highlighting row percentage values in table cells. Default is blue.}

\item{tdcol.col}{Color for highlighting column percentage values in table cells. Default is green.}

\item{emph.total}{logical, if \code{TRUE}, the total column and row will be emphasized with a
different background color. See \code{emph.color}.}

\item{emph.color}{logical, if \code{emph.total = TRUE}, this color value will be used
for painting the background of the total column and row. Default is a light grey.}

\item{prc.sign}{The percentage sign that is printed in the table cells, in HTML-format.
Default is \code{"&nbsp;\%"}, hence the percentage sign has a non-breaking-space after
the percentage value.}

\item{hundret}{Default value that indicates the 100-percent column-sums (since rounding values
may lead to non-exact results). Default is \code{"100.0"}.}
}
\value{
An object of class \code{\link[stats]{ftable}}.
}
\description{
This function creates a labelled flat table or flat
             proportional (marginal) table.
}
\examples{
data(efc)

# flat table with counts
sjt.ftab(efc, e42dep, c172code, e16sex)

# flat table with proportions
sjt.ftab(efc, e42dep, c172code, e16sex, margin = "row")

}

