% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{sjt.glmer}
\alias{sjt.glmer}
\alias{sjt.glm}
\alias{sjt.lmer}
\alias{sjt.lm}
\title{Deprecated functions}
\usage{
sjt.glmer(...)

sjt.glm(...)

sjt.lmer(...)

sjt.lm(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
