% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabGrpmean.R
\name{sjt.grpmean}
\alias{sjt.grpmean}
\title{Summary of grouped means as HTML table}
\usage{
sjt.grpmean(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Computes mean, sd and se for each sub-group.
}
