% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabLinReg.R
\name{sjt.lmer}
\alias{sjt.lmer}
\title{Summary of linear mixed effects models as HTML table}
\usage{
sjt.lmer(..., pred.labels = NULL, depvar.labels = NULL,
  remove.estimates = NULL, group.pred = FALSE, p.numeric = TRUE,
  emph.p = FALSE, p.zero = FALSE, p.kr = TRUE,
  separate.ci.col = TRUE, newline.ci = TRUE, show.est = TRUE,
  show.std = NULL, show.ci = TRUE, show.se = FALSE,
  show.header = FALSE, show.col.header = TRUE, show.r2 = TRUE,
  show.icc = TRUE, show.re.var = TRUE, show.fstat = FALSE,
  show.aic = FALSE, show.aicc = FALSE, show.dev = FALSE,
  string.pred = "Predictors", string.dv = "Dependent Variables",
  string.interc = "(Intercept)", string.obs = "Observations",
  string.est = "B", string.std = "std. Beta", string.ci = "CI",
  string.se = "std. Error", string.p = "p",
  ci.hyphen = "&nbsp;&ndash;&nbsp;", minus.sign = "&#45;",
  digits.est = 2, digits.std = 2, digits.p = 3, digits.ci = 2,
  digits.se = 2, digits.summary = 3, cell.spacing = 0.2,
  cell.gpr.indent = 0.6, sep.column = TRUE, CSS = NULL,
  encoding = NULL, file = NULL, use.viewer = TRUE,
  remove.spaces = TRUE)
}
\arguments{
\item{...}{One or more regression models, including glm's or mixed models.
May also be a \code{list} with fitted models. See 'Examples'.}

\item{pred.labels}{Character vector with labels of predictor variables.
If not \code{NULL}, \code{pred.labels} will be used in the first
table column with the predictors' names. By default, if \code{auto.label = TRUE}
and \code{\link[sjlabelled]{get_term_labels}} is called to retrieve the labels
of the coefficients, which will be used as predictor labels.
If \code{pred.labels = ""} or \code{auto.label = FALSE}, the raw
variable names as used in the model formula are used as predictor
labels. If \code{pred.labels} is a named vector, predictor labels (by
default, the names of the model's coefficients) will be matched with the
names of \code{pred.labels}. This ensures that labels always match the
related predictor in the table, no matter in which way the predictors
are sorted. See 'Examples'.}

\item{p.numeric}{Logical, if \code{TRUE}, the p-values are printed
as numbers. If \code{FALSE} (default), asterisks are used.}

\item{emph.p}{Logical, if \code{TRUE}, significant p-values are shown bold faced.}

\item{show.est}{Logical, if \code{TRUE}, the estimates are printed.}

\item{show.std}{Indicates whether standardized beta-coefficients should
also printed, and if yes, which type of standardization is done.
See 'Details'.}

\item{show.ci}{Either logical, and if \code{TRUE}, the confidence intervals
is printed to the table; if \code{FALSE}, confidence intervals are
omitted. Or numeric, between 0 and 1, indicating the range of the
confidence intervals.}

\item{show.se}{Logical, if \code{TRUE}, the standard errors are also printed.}

\item{show.r2}{Logical, if \code{TRUE}, the r-squared value is also printed.
Depending on the model, these might be pseudo-r-squared values, or Bayesian
r-squared etc. See \code{\link[sjstats]{r2}} for details.}

\item{show.icc}{Logical, if \code{TRUE}, prints the intraclass correlation
coefficient for mixed models. See \code{\link[sjstats]{icc}} for details.}

\item{show.re.var}{Logical, if \code{TRUE}, prints the random effect variances
for mixed models. See \code{\link[sjstats]{re_var}} for details.}

\item{show.fstat}{Logical, if \code{TRUE}, the F-statistics for each model is
printed in the table summary. This option is not supported by all model types.}

\item{show.aic}{Logical, if \code{TRUE}, the AIC value for each model is printed
in the table summary.}

\item{show.aicc}{Logical, if \code{TRUE}, the second-order AIC value for each model
is printed in the table summary.}

\item{show.dev}{Logical, if \code{TRUE}, shows the deviance of the model.}

\item{string.pred}{Character vector,used as headline for the predictor column.
Default is \code{"Predictors"}.}

\item{string.std}{Character vector, used for the column heading of standardized beta coefficients. Default is \code{"std. Beta"}.}

\item{string.ci}{Character vector, used for the column heading of confidence interval values. Default is \code{"CI"}.}

\item{string.se}{Character vector, used for the column heading of standard error values. Default is \code{"std. Error"}.}

\item{string.p}{Character vector, used for the column heading of p values. Default is \code{"p"}.}

\item{ci.hyphen}{Character vector, indicating the hyphen for confidence interval range.
May be an HTML entity. See 'Examples'.}

\item{minus.sign}{string, indicating the minus sign for negative numbers.
May be an HTML entity. See 'Examples'.}

\item{digits.p}{Amount of decimals for p-values}

\item{CSS}{A \code{\link{list}} with user-defined style-sheet-definitions,
according to the \href{http://www.w3.org/Style/CSS/}{official CSS syntax}.
See 'Details' or \href{../doc/table_css.html}{this package-vignette}.}

\item{encoding}{String, indicating the charset encoding used for variable and
value labels. Default is \code{NULL}, so encoding will be auto-detected
depending on your platform (e.g., \code{"UTF-8"} for Unix and \code{"Windows-1252"} for
Windows OS). Change encoding if specific chars are not properly displayed (e.g. German umlauts).}

\item{file}{Destination file, if the output should be saved as file.
If \code{NULL} (default), the output will be saved as temporary file and
openend either in the IDE's viewer pane or the default web browser.}

\item{use.viewer}{Logical, if \code{TRUE}, the HTML table is shown in the IDE's
viewer pane. If \code{FALSE} or no viewer available, the HTML table is
opened in a web browser.}

\item{remove.spaces}{Logical, if \code{TRUE}, leading spaces are removed from all lines in the final string
that contains the html-data. Use this, if you want to remove parantheses for html-tags. The html-source
may look less pretty, but it may help when exporting html-tables to office tools.}
}
\value{
Invisibly returns
         \itemize{
           \item the web page style sheet (\code{page.style}),
           \item the web page content (\code{page.content}),
           \item the complete html-output (\code{page.complete}) and
           \item the html-table with inline-css for use with knitr (\code{knitr})
           }
           for further use.
}
\description{
Summarizes (multiple) fitted linear mixed effects models
               (estimates, std. beta values etc.)  as HTML table,
               or saves them as file. The fitted models may have different
               predictors, e.g. when comparing different stepwise fitted models.
}
\details{
Concerning the \code{show.std} argument, \code{show.std = "std"}
           will print normal standardized estimates. For \code{show.std = "std2"},
           however, standardization of estimates follows
           \href{http://www.stat.columbia.edu/~gelman/research/published/standardizing7.pdf}{Gelman's (2008)}
           suggestion, rescaling the estimates by dividing them by two standard
           deviations instead of just one. Resulting coefficients are then
           directly comparable for untransformed binary predictors. This type
           of standardization uses the \code{\link[arm]{standardize}}-function
           from the \pkg{arm}-package.
           For backward compatibility reasons, \code{show.std} also may be
           a logical value; if \code{TRUE}, normal standardized estimates are
           printed (same effect as \code{show.std = "std"}). Use
           \code{show.std = NULL} (default) or \code{show.std = FALSE},
           if standardized estimats should not be printed.
           \cr \cr
           Computation of p-values (if necessary and if \code{p.kr = TRUE}) are based
           on conditional F-tests with Kenward-Roger approximation for the df, using
           the \pkg{pbkrtest}-package. If \pkg{pbkrtest} is not available or
           \code{p.kr = FALSE}, computation of p-values is based
           on normal-distribution assumption, treating the t-statistics as Wald
           z-statistics. See 'Details' in \code{\link[sjstats]{p_value}}.
           \cr \cr
           The confidence intervals stem from \pkg{broom}'s
           \code{\link[broom]{tidy}}-function. For linear mixed models, the computation
           method is "Wald" (\code{lme4::confint.merMod(fit, method = "Wald")}).
}
\note{
The variance components of the random effects (see \code{show.re.var}) are
        denoted like:
        \itemize{
         \item within-group variance: sigma-squared
         \item between-group-variance: tau-zero-zero
         \item random-slope-intercept-correlation: rho-zero-one
         }
}
