\name{sju.cronbach}
\alias{sju.cronbach}
\title{Calculates Cronbach's Alpha for a matrix}
\usage{
sju.cronbach(df)
}
\arguments{
  \item{df}{A data frame or matrix with more than 2
  columns.}
}
\value{
The Cronbach's alpha value for \code{df}.
}
\description{
This function calcates the Cronbach's alpha value for each
column of a data frame or matrix.
}
\note{
For use case, see \code{\link{sjp.pca}} and
\code{\link{sjt.pca}}.
}

