% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjRecode.R
\name{sju.recode}
\alias{sju.recode}
\title{Recode variable values.}
\usage{
sju.recode(...)
}
\arguments{
\item{...}{parameters, see \code{\link{recode}} function from the \code{car} package.}
}
\value{
A variable with recoded values.
}
\description{
Recodes the categories of a variables. Wrapper function that calls
               the \code{\link{recode}} function from the \code{car} package.
}
\examples{
data(efc)
table(efc$e42dep)
table(sju.recode(efc$e42dep, "1:2=1;3:4=2"))
}

