% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjStatistics.R
\name{std_e}
\alias{std_e}
\title{Compute standard error for variables}
\usage{
std_e(x)
}
\arguments{
\item{x}{a (numeric) vector / variable.}
}
\value{
The standard error of variable \code{x}.
}
\description{
Compute standard error for variables
}
\examples{
std_e(rnorm(n = 100, mean = 3))
}

