% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_print.R
\name{tab_df}
\alias{tab_df}
\alias{tab_dfs}
\title{Print data frames as HTML table.}
\usage{
tab_df(x, title = NULL, footnote = NULL, col.header = NULL,
  show.type = FALSE, show.rownames = TRUE, show.footnote = FALSE,
  alternate.rows = FALSE, encoding = "UTF-8", CSS = NULL, file = NULL,
  use.viewer = TRUE, ...)

tab_dfs(x, titles = NULL, footnotes = NULL, col.header = NULL,
  show.type = FALSE, show.rownames = TRUE, show.footnote = FALSE,
  alternate.rows = FALSE, encoding = "UTF-8", CSS = NULL, file = NULL,
  use.viewer = TRUE, ...)
}
\arguments{
\item{x}{For \code{tab_df()}, a data frame; and for \code{tab_dfs()}, a
list of data frames.}

\item{title, titles}{Character vector (of length 1) with table caption. Only
applies to \code{tab_df()}. For \code{tab_dfs()}, add a character vector
as \code{title}-attribute to each data frame, which will then be used as
caption for each table.}

\item{footnote, footnotes}{Character vector (of length 1) with table footnote. Only
applies to \code{tab_df()}. For \code{tab_dfs()}, add a character vector
as \code{footnote}-attribute to each data frame, which will then be used
as footnote under each table.}

\item{col.header}{Character vector with elements used as column header for
the table. If \code{NULL}, column names from \code{x} are used as
column header.}

\item{alternate.rows}{}

\item{encoding}{Character vector, indicating the charset encoding used
for variable and value labels. Default is \code{"UTF-8"}. For Windows
Systems, \code{encoding = "Windows-1252"} might be necessary for proper
display of special characters.}

\item{CSS}{A \code{\link{list}} with user-defined style-sheet-definitions, according to the
\href{http://www.w3.org/Style/CSS/}{official CSS syntax}. For more details,
see \href{../doc/sjtbasic.html}{this package-vignette}, or 'Details' in
\code{\link{sjt.frq}}.}

\item{file}{Destination file, if the output should be saved as file.
If \code{NULL} (default), the output will be saved as temporary file and
openend either in the IDE's viewer pane or the default web browser.}

\item{use.viewer}{Logical, if \code{TRUE}, the HTML table is shown in the IDE's viewer pane. If
\code{FALSE} or no viewer available, the HTML table is opened in a web browser.}
}
\value{
A list with following items:
  \itemize{
    \item the web page style sheet (\code{page.style}),
    \item the HTML content of the data frame (\code{page.content}),
    \item the complete HTML page, including header, style sheet and body (\code{page.complete})
    \item the HTML table with inline-css for use with knitr (\code{knitr})
    \item the file path, if the HTML page should be saved to disk (\code{file})
  }
}
\description{
These functions print data frames as HTML-table, showing
  the results in RStudio's viewer pane or in a web browser.
}
