% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{to_label}
\alias{to_label}
\title{Replaces variable values with their associated value labels}
\usage{
to_label(variable)
}
\arguments{
\item{variable}{A (factor) variable.}
}
\value{
A factor variable containing with the replaced value labels.
}
\description{
This function converts (replaces) variable values (of factors) with their
associated value labels. Might be helpful for factor variables.
For instance, if you have a Gender variable with 0/1, and associated
labels are male/female, this function would convert all 0 to male and
all 1 to female in the data frame.
}
\examples{
data(efc)
print(get_val_labels(efc)['c161sex'])
head(efc$c161sex)
head(to_label(efc$c161sex))

print(get_val_labels(efc)['e42dep'])
table(efc$e42dep)
table(to_label(efc$e42dep))
}
\seealso{
\itemize{
           \item \code{\link{to_fac}}
           \item \code{\link{to_value}}
           \item \code{\link{get_val_labels}}
           \item \code{\link{get_var_labels}}
           \item \code{\link{read_spss}}
           }
}

