% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{to_sjPlot}
\alias{to_sjPlot}
\title{Convert a haven-imported data frame to sjPlot format}
\usage{
to_sjPlot(x)
}
\arguments{
\item{x}{a data frame, which contains \code{labelled} class vectors}
}
\value{
a data frame with 'sjPlot' friendly vector classes and attached
          label attributes.
}
\description{
This function converts a data frame, which was imported with any of
               \code{haven}'s read functions and contains \code{labelled} class
               vectors, into an sjPlot friendly data frame format.
}

