% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{to_value}
\alias{to_value}
\title{Converts factors to numeric variables}
\usage{
to_value(fac, startAt = 1)
}
\arguments{
\item{fac}{A (factor) variable.}

\item{startAt}{the starting index, i.e. numeric value of the variable.}
}
\value{
A numeric variable with values ranging from \code{startAt} to
          \code{startAt} + length of factor levels.
}
\description{
This function converts (replaces) factor values with the
related factor level index number, thus the factor is converted to
a numeric variable.
}
\examples{
data(efc)
test <- to_label(efc$e42dep)
table(test)

table(to_value(test))
hist(to_value(test, 0))
}
\seealso{
\itemize{
           \item \code{\link{to_fac}}
           \item \code{\link{to_label}}
           \item \code{\link{get_val_labels}}
           \item \code{\link{get_var_labels}}
           \item \code{\link{read_spss}}
           }
}

