% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{write_spss}
\alias{write_spss}
\title{Write content of data frame to SPSS sav-file}
\usage{
write_spss(x, path)
}
\arguments{
\item{x}{data frame that should be saved as SPSS sav-file.}

\item{path}{file path to the SPSS dataset.}
}
\description{
This function saves the content of a data frame to an SPSS sav-file.
}
\note{
You don't need to take care whether variables have been imported with
        the \code{\link{read_spss}} function from this package or from \code{haven}
        or even the \code{foreign} package, or if you have imported SPSS data and
        created new variables. This function does all necessary data preparation
        to write a properly labelled SPSS sav file.
}
\seealso{
\itemize{
           \item \href{http://www.strengejacke.de/sjPlot/datainit/}{sjPlot manual: data initialization}
           \item \href{http://www.strengejacke.de/sjPlot/view_spss/}{sjPlot manual: inspecting (SPSS imported) data frames}
           \item \code{\link{get_val_labels}}
           \item \code{\link{get_var_labels}}
           \item \code{\link{to_label}}
           \item \code{\link{to_value}}
           \item \code{\link{to_fac}}
           \item \code{\link{view_spss}}
           \item \code{\link{read_spss}}
           }
}

