% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjRecode.R
\name{adjust_plot_range}
\alias{adjust_plot_range}
\title{Adjust y range of ggplot-objects}
\usage{
adjust_plot_range(gp, upperMargin = 1.05)
}
\arguments{
\item{gp}{A ggplot-object. Usually, this will be returned by most of this
package's plotting functions.}

\item{upperMargin}{Defines the margin of the upper y bound of the plot. This value will
be multiplied with the total y range. Default is 1.05, which means that the upper
margin of the plot is about 5 percent of the "visible" plot area (i.e. the y-range
is 105 percent of the actual needed range to make all object visible).}
}
\value{
The same ggplot-object, with adjusted y-range, so all graphics and labels
         should be visible.
}
\description{
This method adjusts the y-range of a ggplot-object, which is useful when
               value labels are outside of the plot region. A modified ggplot-object will
               be returned with adjusted y-range so everything should be visible.
               Note that this function only works on \code{scale_y_continuous}.
}
\note{
Note that this function only works on \code{scale_y_continuous}.
}
\examples{
# sample data set
data(efc)
# show frequencies of relationship-variable and
# retrieve plot object
gp <- sjp.frq(efc$e15relat, printPlot=FALSE)
# show current plot
plot(gp$plot)
# show adjusted plot
adjust_plot_range(gp$plot)
}
\references{
\href{http://www.r-bloggers.com/setting-axis-limits-on-ggplot-charts/}{r-bloggers.com}
}

