% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjStatistics.R
\name{cv}
\alias{cv}
\title{Compute coefficient of variance}
\usage{
cv(x)
}
\arguments{
\item{x}{a (numeric) vector / variable.}
}
\value{
The coefficient of variance (standard deviation divided
          by mean) of \code{x}.
}
\description{
Compute coefficient of variance (standard deviation divided
               by mean).
}
\examples{
data(efc)
cv(efc$e17age)
}

