% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Various plot types for regression models}
\usage{
plot_model(model, type = c("est", "re", "eff", "pred", "int", "std", "std2",
  "slope", "resid"), exponentiate, terms = NULL, sort.est = FALSE,
  rm.terms = NULL, group.terms = NULL, pred.type = c("fe", "re"),
  title = NULL, axis.title = NULL, axis.labels = NULL, axis.lim = NULL,
  grid.breaks = NULL, ci.lvl = NULL, show.intercept = FALSE,
  show.values = FALSE, show.p = TRUE, show.data = FALSE,
  value.offset = NULL, dot.size = NULL, line.size = NULL,
  colors = "Set1", facets, wrap.title = 50, wrap.labels = 25,
  case = NULL, digits = 2, vline.type = 2, vline.color = "grey70", ...)
}
\arguments{
\item{type}{Type of plot. Use one of following:
\describe{
  \item{\code{"est"}}{(default) for forest-plot of estimates. If the fitted model only contains one predictor, slope-line is plotted.}
  \item{\code{"pred"}}{to plot predicted values (marginal effects) for specific model terms. See 'Details'.}
  \item{\code{"eff"}}{to plot marginal effects of all terms in \code{model}. Note that interaction terms are excluded from this plot.}
  \item{\code{"int"}}{to plot marginal effects of interaction terms in \code{model}.}
  \item{\code{"std"}}{for forest-plot of standardized beta values.}
  \item{\code{"std2"}}{for forest-plot of standardized beta values, however, standardization is done by dividing by two sd (see 'Details').}
  \item{\code{"slope"}}{to plot regression lines for each single predictor of the fitted model, against the response (linear relationship between each model term and response).}
  \item{\code{"resid"}}{to plot regression lines for each single predictor of the fitted model, against the residuals (linear relationship between each model term and residuals). May be used for model diagnostics.}
  \item{\code{"ma"}}{to check model assumptions.}
  \item{\code{"vif"}}{to plot Variance Inflation Factors.}
}}

\item{terms}{Character vector with the names of those from \code{model}, which
should be used to plot for. This argument depends on the plot-type;
for \code{type = "pred"} or \code{type = "eff"}, \code{terms} indicates
for which terms marginal effects should be displayed. At least one term
is required to calculate effects, maximum length is three terms,
where the second and third term indicate the groups, i.e. predictions
of first term are grouped by the levels of the second (and third)
term. Indicating levels in square brackets allows for selecting
only specific groups. Term name and levels in brackets must be
separated by a whitespace character, e.g.
\code{terms = c("age", "education [1,3]")}. For more details, see
\code{\link[ggeffects]{ggpredict}}.}

\item{...}{Other arguments, passed down to various functions. Here is the
description of these arguments in detail.
\describe{
  \item{\strong{stanreg-models (rstanarm)}}{
    For \code{stanreg}-models (fitted with the \pkg{rstanarm}-package)
    and plot-type \code{type = "est"}, you can specify numeric values
    between 0 and 1 for \code{prob.inner} and \code{prob.outer}, which
    will then be used as inner and outer probabilities for the uncertainty
    intervals (HDI). By default, the inner probability is 0.5 and the
    outer probability is 0.89 (unless \code{ci.lvl} is specified - in
    this case, \code{ci.lvl} is used as outer probability). The
    Bayesian point estimate is, by default, the median of the posterior
    distribution. Use \code{bpe} to define other functions to calculate
    the Bayesion point estimate. \code{bpe} needs to be a character naming
    the specific function, which is passed to the \code{fun}-argument
    in \code{\link[sjstats]{typical_value}}. So, \code{bpe = "mean"}
    would calculate the mean value of the posterior distribution.
  }
}}
}
\description{
To Do...
}
