% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjPlotGLME.R
\name{sjp.glmer}
\alias{sjp.glmer}
\title{Plot estimates, predictions or effects of generalized linear mixed effects models}
\usage{
sjp.glmer(fit, type = "re", vars = NULL, ri.nr = NULL, emph.grp = NULL,
  title = NULL, geom.size = NULL, geom.colors = "Set1",
  hideErrorBars = FALSE, show.intercept = TRUE,
  stringIntercept = "(Intercept)", sort.est = NULL, var.labels = NULL,
  axisTitle.x = NULL, axisTitle.y = NULL, ylim = NULL,
  facet.grid = TRUE, free.scale = FALSE, vline.type = 2,
  vline.color = "grey70", remove.estimates = NULL, show.values = TRUE,
  labelDigits = 2, y.offset = 0.1, showPValueLabels = TRUE,
  fade.ns = FALSE, show.ci = FALSE, sample.n = NULL,
  show.legend = FALSE, printPlot = TRUE)
}
\arguments{
\item{fit}{a fitted model as returned by the \code{\link[lme4]{glmer}}-function.}

\item{type}{type of plot. Use one of following:
\describe{
  \item{\code{"re"}}{(default) for odds or incidents ratios of random effects}
  \item{\code{"fe"}}{for odds or incidents ratios of fixed effects}
  \item{\code{"fe.cor"}}{for correlation matrix of fixed effects}
  \item{\code{"re.qq"}}{for a QQ-plot of random effects (random effects quantiles against standard normal quantiles)}
  \item{\code{"fe.slope"}}{to plot probability or incidents curves (predicted probabilities or incidents) of all fixed effects coefficients. Use \code{facet.grid} to decide whether to plot each coefficient as separate plot or as integrated faceted plot. See 'Details'.}
  \item{\code{"ri.slope"}}{to plot probability or incidents curves (predicted probabilities or incidents) of random intercept variances for all fixed effects coefficients. Use \code{facet.grid} to decide whether to plot each coefficient as separate plot or as integrated faceted plot. See 'Details'.}
  \item{\code{"rs.ri"}}{for fitted probability curves (predicted probabilities) indicating the random slope-intercept pairs. Use this to visualize the random parts of random slope-intercept (or repeated measure) models. When having too many groups, use \code{sample.n} argument.}
  \item{\code{"eff"}}{to plot marginal effects of predicted probabilities or incidents for each fixed term, where remaining co-variates are set to the mean. Use \code{facet.grid} to decide whether to plot each coefficient as separate plot or as integrated faceted plot. See 'Details'.}
  \item{\code{"pred"}}{to plot predicted probabilities or incidents for the response, related to specific model predictors and conditioned on random effects. See 'Details'.}
  \item{\code{"pred.fe"}}{to plot predicted probabilities or incidents for the response, related to specific model predictors, only for fixed effects. See 'Details'.}
  \item{\code{"ma"}}{to check model assumptions. Note that only argument \code{fit} applies to this plot type. All other arguments are ignored.}
}}

\item{vars}{numeric vector with column indices of selected variables or a character vector with
variable names of selected variables from the fitted model, which should be used to plot
estimates, fixed effects slopes (for \code{\link[lme4]{lmer}}) or probability or
incidents curves
(for \code{\link[lme4]{glmer}}) of random intercepts.}

\item{ri.nr}{numeric vector. If \code{type = "re"} or \code{type = "ri.slope"},
and fitted model has more than one random intercept, \code{ri.nr} indicates
which random effects of which random intercept (or: which list elements
of \code{\link[lme4]{ranef}}) will be plotted. Default is \code{NULL},
so all random effects will be plotted.}

\item{emph.grp}{numeric vector with index numbers of grouping levels (from random effect).
If \code{type = "ri.slope"} and \code{facet.grid = FALSE},
an integrated plot of predicted probabilities of fixed effects resp. fixed
effects slopes for each grouping level is plotted. To better find
certain groups, use this argument to emphasize these groups in the plot.
See 'Examples'.}

\item{title}{character vector with one or more labels that are used as plot title. If
\code{type = "re"}, use the predictors' variable labels as titles.}

\item{geom.size}{size of geoms (point size or line size, depending on \code{type}-argument).}

\item{geom.colors}{user defined color palette for geoms. Must either be vector with two color values
or a specific color palette code. See 'Note' in \code{\link{sjp.grpfrq}}.}

\item{hideErrorBars}{logical, if \code{TRUE}, the error bars that indicate the confidence intervals of the odds ratios are not
shown.}

\item{show.intercept}{logical, if \code{TRUE}, the intercept of the fitted model is also plotted.
Default is \code{FALSE}. For \code{glm}'s, please note that due to exponential 
transformation of estimates, the intercept in some cases can not be calculated, thus the
function call is interrupted and no plot printed.}

\item{stringIntercept}{string, label of intercept estimate on the y axis. Only applies, if \code{show.intercept}
is \code{TRUE} and \code{var.labels} is not \code{NULL}.}

\item{sort.est}{determines in which way estimates are sorted in the plot:
\itemize{
  \item If \code{NULL} (default), no sorting is done and estimates are sorted in order of model coefficients.
  \item If \code{sort.est = "sort.all"}, estimates are re-sorted for each coefficient (only applies if \code{type = "re"} and \code{facet.grid = FALSE}), i.e. the estimates of the random effects for each predictor are sorted and plotted to an own plot.
  \item If \code{type = "fe"} or \code{type = "fe.std"}, \code{TRUE} will sort estimates
  \item Else, specify a predictor's / coefficient's name to sort estimates according to this coefficient.
  }
  See 'Examples'.}

\item{var.labels}{character vector with labels for the model terms, used as axis labels.
For mixed models, should either be vector of fixed effects variable labels 
(if \code{type = "fe"} or \code{type = "fe.std"}) or a vector of group (value)
labels from the random intercept's categories (if \code{type = "re"}).}

\item{axisTitle.x}{title for the x axis. If not specified, a default labelling depending
on the plot type is chosen.}

\item{axisTitle.y}{title for the y axis. If not specified, a default labelling depending
on the plot type is chosen.}

\item{ylim}{numeric vector of length two, defining lower and upper axis limits
of the y scale. By default, this argument is set to \code{NULL}, i.e. the 
y-axis fits to the required range of the data.}

\item{facet.grid}{\code{TRUE} when each plot should be plotted separately instead of
an integrated (faceted) single graph.}

\item{free.scale}{if \code{TRUE} and \code{facet.grid = TRUE}, each facet grid
gets its own fitted scale. If \code{free.scale = FALSE}, each facet in
the grid has the same scale range.}

\item{vline.type}{linetype of the vertical "zero point" line. Default is \code{2} (dashed line).}

\item{vline.color}{color of the vertical "zero point" line. Default value is \code{"grey70"}.}

\item{remove.estimates}{character vector with coefficient names that indicate
which estimates should be removed from the plot.
\code{remove.estimates = "est_name"} would remove the estimate \emph{est_name}. Default
is \code{NULL}, i.e. all estimates are printed.}

\item{show.values}{logical, whether values should be plotted or not.}

\item{labelDigits}{numeric, amount of digits for rounding the estimates (see \code{show.values}).
Default is 2, i.e. estimates have 2 digits after decimal point.}

\item{y.offset}{numeric, offset for text labels when their alignment is adjusted 
to the top/bottom of the geom (see \code{hjust} and \code{vjust}).}

\item{showPValueLabels}{logical, whether the significance levels of each coefficient should be appended
to values or not.}

\item{fade.ns}{if \code{TRUE}, non significant estimates will be printed in slightly faded colors.}

\item{show.ci}{logical, if \code{TRUE}, depending on \code{type}, a condifence
region is added to the plot.}

\item{sample.n}{numeric vector. only applies, if \code{type = "rs.ri"}. If
plot has many random intercepts (grouping levels), overplotting of
regression lines may occur. In this case, consider random sampling of
grouping levels. If \code{sample.n} is of length 1, a random sample
of \code{sample.n} observation is selected to plot random intercepts.
If \code{sample.n} is of length > 1, random effects indicated by
the values in \code{sample.n} are selected to plot random effects.
Use the latter option to always select a fixed, identical set of
random effects for plotting (useful when ecomparing multiple models).}

\item{show.legend}{logical, if \code{TRUE}, for mixed models
\code{type = "rs.ri"}, a legend for group levels of
the random intercept is shown. For \code{lm} and \code{glm},
a legend for grouped estimates is shown.}

\item{printPlot}{logical, if \code{TRUE} (default), plots the results as graph. Use \code{FALSE} if you don't
want to plot any graphs. In either case, the ggplot-object will be returned as value.}
}
\value{
(Insisibily) returns, depending on the plot type
         \itemize{
           \item The ggplot-object (\code{plot}). For multiple plots and if \code{facet.grid = FALSE}) a \code{plot.list} is returned.
           \item A data frame \code{data} with the data used to build the ggplot-object(s), or a list of data frames (\code{data.list}).
           }
}
\description{
By default, this function plots estimates (odds, risk or incidents
               ratios, i.e. exponentiated coefficients, depending on family and
               link function)
               with confidence intervals of either fixed effects or random effects of
               generalized linear mixed effects models (that have been fitted with the
               \code{\link[lme4]{glmer}}-function of the \pkg{lme4}-package).
               Furthermore, this function also plots predicted probabilities /
               incidents or diagnostic plots.
}
\details{
\describe{
           \item{\code{type = "fe.slope"}}{the predicted values
           are based on the fixed effects intercept's estimate and each specific
           fixed term's estimate. All other fixed effects are set to zero (i.e. ignored),
           which corresponds to \code{family(fit)$linkinv(eta = b0 + bi * xi)} (where \code{xi}
           is the estimate of fixed effects and \code{b0} is the intercept of
           the fixed effects; the inverse link-function is used). This plot type 
           may give similar results as \code{type = "pred"}, however, \code{type = "fe.slope"} 
           does not adjust for other predictors.}
           \item{\code{type = "eff"}}{plots the marginal effects of model predictors.
           Unlike \code{type = "fe.slope"}, the predicted values computed by 
           \code{type = "eff"} are adjusted for all co-variates, which are
           set to the mean, as returned by the \code{\link[effects]{allEffects}} function.}
           \item{\code{type = "ri.slope"}}{the predicted values
           are based on the fixed effects intercept, plus each random intercept and
           each specific  fixed term's estimate. All other fixed effects are set to zero (i.e. ignored),
           which corresponds to \code{family(fit)$linkinv(eta = b0 + b0[r1-rn] + bi * xi)}
           (where \code{xi} is the estimate of fixed effects, \code{b0} is the intercept of
           the fixed effects and \code{b0[r1-rn]} are all random intercepts).}
           \item{\code{type = "rs.ri"}}{the predicted values are based
           on the fixed effects intercept, plus each random intercept and
           random slope. This plot type is intended to plot the random part, i.e.
           the predicted probabilities or incident rates of each random slope
           for each random intercept.
           Since the random intercept specifies the deviance from the gloabl
           intercept, the global intercept is always included. In case of overplotting,
           use the \code{sample.n} argument to randomly sample a limited amount
           of groups.}
           \item{\code{type = "coef"}}{forest plot of joint fixed and random
           effect coefficients, as retrieved by \code{\link[lme4]{coef.merMod}},
           it's simply \code{\link[lme4]{ranef} + \link[lme4]{fixef}}.}
           \item{\code{type = "pred"} or \code{type = "pred.fe"}}{predicted 
           values against response, only fixed effects or
           conditional on random intercept. It's calling
           \code{predict(fit, type = "response", re.form = NA)} resp.
           \code{predict(fit, type = "response", re.form = NULL)} to
           compute the values.
           This plot type requires the \code{vars} argument to select specific terms
           that should be used for the x-axis and - optional - as grouping factor. 
           Hence, \code{vars} must be a character vector with the names of
           one or two model predictors. See 'Examples'.}
         }
}
\note{
\itemize{
         \item{Computation of p-values (if necessary) are based on
               Wald chi-square tests from the \code{Anova}-function of the \pkg{car}-package.}
         \item{Plot types use the inverse link-function to calculate predicted
               probabilites or incidents rates. Thus, this function should work
               with different model families and link functions; however, the
               plot or axis title may not use the exact terminology regarding
               model family or link function.}
         \item{Thanks go to Robert Reijntjes from
               Leiden University Medical Center for sharing R code that is used
               to compute fixed effects correlation matrices and qq-plots of
               random effects.}
       }
}
\examples{
library(lme4)
library(sjmisc)
# create binary response
sleepstudy$Reaction.dicho <- dicho(sleepstudy$Reaction, dich.by = "median")
# fit model
fit <- glmer(Reaction.dicho ~ Days + (Days | Subject),
             sleepstudy,
             family = binomial("logit"))

# simple plot
sjp.glmer(fit)

# sort by predictor Days
sjp.glmer(fit, sort.est = "Days")

data(efc)
# create binary response
efc$hi_qol <- dicho(efc$quol_5)
# prepare group variable
efc$grp = as.factor(efc$e15relat)
levels(x = efc$grp) <- get_labels(efc$e15relat)
# data frame for fitted model
mydf <- data.frame(hi_qol = as.factor(efc$hi_qol),
                   sex = as.factor(efc$c161sex),
                   c12hour = as.numeric(efc$c12hour),
                   neg_c_7 = as.numeric(efc$neg_c_7),
                   grp = efc$grp)
# fit glmer
fit <- glmer(hi_qol ~ sex + c12hour + neg_c_7 + (1|grp),
             data = mydf,
             family = binomial("logit"))

# plot and sort fixed effects
sjp.glmer(fit,
          type = "fe",
          sort.est = TRUE)

# plot probability curves (predicted probabilities)
# for each covariate, grouped by random intercepts
# in integrated plots, emphasizing groups 1 and 4
sjp.glmer(fit,
          type = "ri.slope",
          emph.grp = c(1, 4),
          facet.grid = FALSE)

# plot probability curve (predicted probabilities)
# of fixed effect, only for coefficient "neg_c_7"
sjp.glmer(fit,
          type = "fe.slope",
          vars = "neg_c_7")

# plot predicted probabilities for response,
# related to model predictor, conditioned on random effects
sjp.glmer(fit,
          type = "pred",
          vars = "neg_c_7")

# plot predicted probabilities for response,
# related to model predictor, grouped
sjp.glmer(fit,
          type = "pred.fe",
          vars = c("neg_c_7", "sex"))
          
# non faceted, with ci           
sjp.glmer(fit, type = "pred.fe", vars = c("neg_c_7", "sex"), 
          show.ci = TRUE, facet.grid = FALSE)

                     
}
\seealso{
\href{http://www.strengejacke.de/sjPlot/sjp.glmer/}{sjPlot manual: sjp.glmer}
}

