% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sjp.glmm}
\alias{sjp.glmm}
\title{Plot odds ratios (forest plots) of multiple fitted glm's}
\usage{
sjp.glmm(..., title = NULL, labelDependentVariables = NULL,
  legendDepVarTitle = "Dependent Variables", legendPValTitle = "p-level",
  stringModel = "Model", axisLabels.y = NULL, axisTitle.x = "Odds Ratios",
  axisLimits = NULL, breakTitleAt = 50, breakLabelsAt = 25,
  breakLegendAt = 20, gridBreaksAt = 0.5, transformTicks = FALSE,
  geom.size = 3, geom.spacing = 0.4, geom.colors = "Dark2", nsAlpha = 1,
  usePShapes = FALSE, interceptLineType = 2,
  interceptLineColor = "grey70", coord.flip = TRUE, showIntercept = FALSE,
  showAxisLabels.y = TRUE, showValueLabels = TRUE, labelDigits = 2,
  showPValueLabels = TRUE, hideLegend = FALSE, facet.grid = FALSE,
  printPlot = TRUE)
}
\arguments{
\item{...}{One or more fitted glm-objects.}

\item{title}{Diagram's title as string.
Example: \code{title=c("my title")}}

\item{labelDependentVariables}{Labels of the dependent variables of all fitted models
which have been used as first parameter(s), provided as char vector.}

\item{legendDepVarTitle}{A character vector used for the title of the dependent variable's legend.
Default is \code{"Dependent Variables"}.}

\item{legendPValTitle}{A character vector used for the title of the significance level's legend.
Default is \code{"p-level"}. Only applies if \code{usePShapes} is \code{TRUE}.}

\item{stringModel}{String constant used as legend text for the model names in case no
labels for the dependent variables are provided (see \code{labelDependentVariables}).
Default is \code{"Model"}.}

\item{axisLabels.y}{Labels of the predictor variables (independent vars, odds) that are used for labelling the
axis. Passed as vector of strings.
Example: \code{axisLabels.y=c("Label1", "Label2", "Label3")}
Note: If you use the \code{\link{sji.SPSS}} function and the \code{\link{sji.getValueLabels}} function, you receive a
\code{list} object with label strings. The labels may also be passed as list object. They will be unlisted and
converted to character vector automatically.}

\item{axisTitle.x}{A label ("title") for the x axis.}

\item{axisLimits}{Defines the range of the axis where the beta coefficients and their confidence intervalls
are drawn. By default, the limits range from the lowest confidence interval to the highest one, so
the diagram has maximum zoom. Use your own values as 2-value-vector, for instance: \code{limits=c(-0.8,0.8)}.}

\item{breakTitleAt}{Wordwrap for diagram title. Determines how many chars of the title are displayed in
one line and when a line break is inserted into the title}

\item{breakLabelsAt}{Wordwrap for diagram labels. Determines how many chars of the category labels are displayed in
one line and when a line break is inserted}

\item{breakLegendAt}{Wordwrap for legend, i.e. names of the dependent variables of each fitted model.
See parameter \code{labelDependentVariables}. Determines how many chars of each dependent variable name
is displayed in one line in the legend and when a line break is inserted}

\item{gridBreaksAt}{Sets the breaks on the y axis, i.e. at every n'th position a major
grid is being printed. Default is 0.5}

\item{transformTicks}{if \code{TRUE}, the grid bars have exponential distances, i.e. they
visually have the same distance from one grid bar to the next. Default is \code{FALSE} which
means that grids are plotted on every \code{gridBreaksAt}'s position, thus the grid bars
become narrower with higher odds ratio values.}

\item{geom.size}{The size of the points that indicate the odds ratios. Default is 3.}

\item{geom.spacing}{Defines the space between the dots and error bars of the plotted fitted models. Default
is 0.3.}

\item{geom.colors}{A vector with colors for representing the odds ratio values (i.e. points and error bars)
of the different fitted models. Thus, the length of this vector must be equal to
the length of supplied fitted models, so each model is represented by its own color.
You can use:
\itemize{
  \item If not specified, the qualitative \code{"Dark2"} color brewer palette will be used.
  \item If \code{"gs"}, a greyscale will be used.
  \item If \code{geom.colors} is any valid color brewer palette name, the related \href{http://colorbrewer2.org}{color brewer} palette will be used. Use \code{display.brewer.all()} from the \code{RColorBrewer} package to view all available palette names.
  }
Else specify your own color values as vector (e.g. \code{geom.colors=c("#f00000", "#00ff00", "#0080ff")}).}

\item{nsAlpha}{The alpha level (transparancy) of non significant predicors. Points and error bars
are affected by this value and plotted with a slight transparancy. Default is 1.}

\item{usePShapes}{If \code{TRUE}, significant levels are distinguished by different point shapes and a related
legend is plotted. Default is \code{FALSE}.}

\item{interceptLineType}{The linetype of the intercept line (zero point). Default is \code{2} (dashed line).}

\item{interceptLineColor}{The color of the intercept line. Default value is \code{"grey70"}.}

\item{coord.flip}{If \code{TRUE} (default), predictors are plotted on the left y-axis and estimate
values are plotted on the x-axis.}

\item{showIntercept}{If \code{TRUE}, the intercept of the fitted model is also plotted.
Default is \code{FALSE}. Please note that due to exp-transformation of
estimates, the intercept in some cases can not be calculated, thus the
function call is interrupted and no plot printed.}

\item{showAxisLabels.y}{Whether odds names (predictor labels) should be shown or not.}

\item{showValueLabels}{Whether the beta and standardized beta values should be plotted
to each dot or not.}

\item{labelDigits}{The amount of digits for rounding the estimations (see \code{showValueLabels}).
Default is 2, i.e. estimators have 2 digits after decimal point.}

\item{showPValueLabels}{Whether the significance levels of each coefficient should be appended
to values or not.}

\item{hideLegend}{Indicates whether legend (guide) should be shown or not.}

\item{facet.grid}{\code{TRUE} when each model should be plotted as single facet instead of
an integrated single graph.}

\item{printPlot}{If \code{TRUE} (default), plots the results as graph. Use \code{FALSE} if you don't
want to plot any graphs. In either case, the ggplot-object will be returned as value.}
}
\value{
(Insisibily) returns the ggplot-object with the complete plot (\code{plot}) as well as the data frame that
          was used for setting up the ggplot-object (\code{df}).
}
\description{
Plot odds ratios (forest plots) of multiple fitted glm's with confidence intervalls in one plot.
}
\examples{
# prepare dummy variables for binary logistic regression
y1 <- ifelse(swiss$Fertility<median(swiss$Fertility), 0, 1)
y2 <- ifelse(swiss$Infant.Mortality<median(swiss$Infant.Mortality), 0, 1)
y3 <- ifelse(swiss$Agriculture<median(swiss$Agriculture), 0, 1)

# Now fit the models. Note that all models share the same predictors
# and only differ in their dependent variable (y1, y2 and y3)
fitOR1 <- glm(y1 ~ swiss$Education+swiss$Examination+swiss$Catholic,
              family=binomial(link="logit"))
fitOR2 <- glm(y2 ~ swiss$Education+swiss$Examination+swiss$Catholic,
              family=binomial(link="logit"))
fitOR3 <- glm(y3 ~ swiss$Education+swiss$Examination+swiss$Catholic,
              family=binomial(link="logit"))

# plot multiple models
sjp.glmm(fitOR1, fitOR2, fitOR3, facet.grid=TRUE)

# plot multiple models with legend labels and point shapes instead of value  labels
sjp.glmm(fitOR1, fitOR2, fitOR3,
         labelDependentVariables=c("Fertility", "Infant Mortality", "Agriculture"),
         showValueLabels=FALSE,
         showPValueLabels=FALSE,
         usePShapes=TRUE,
         nsAlpha=0.2)
}
\seealso{
\itemize{
           \item \code{\link{sjp.glm}}
           \item \code{\link{sjt.glm}}
           \item \code{\link{sjp.glm.ma}}
           \item \code{\link{sjp.lmm}}
           }
}

