\encoding{UTF-8} 
\name{sjPlot-package}
\alias{sjPlot-package}
\alias{sjPlot}
\docType{package}
\title{
sjPlot - data visualization for statistics in social science
}
\description{
Collection of several plotting functions for data visualization using ggplot. Results of several statistical analyses (that are commonly used in social sciences) can be visualized using this package, including simple and cross tabulated frequencies, histograms, box plots, (generalized) linear models (forest plots), PCA, correlations etc. \cr \cr
Furthermore, this package contains some tools that are useful when carrying out data
analysis or interpreting data, for instance data set import, variable recoding,
determination of cluster groups, interpretation of interactions in linear models etc. \cr \cr
The functions follow specific naming conventions: \cr
sjc - cluster analysis functions \cr
sji - data import and manipulation functions \cr
sjp - plotting functions \cr
sjt - (HTML) table output functions \cr
sju - statistical utility functions \cr
}
\details{
\tabular{ll}{
Package: \tab sjPlot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-28\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
\url{http://strengejacke.wordpress.com/sjplot-r-package/}
}
