% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjplot.R
\name{sjplot}
\alias{sjplot}
\alias{sjtab}
\title{Wrapper to create plots and tables within a pipe-workflow}
\usage{
sjplot(.data, ..., fun = c("frq", "grpfrq", "xtab", "gpt", "scatter"))

sjtab(.data, ..., fun = c("frq", "xtab"))
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Names of variables that should be plotted, and also further 
arguments passed down to the \pkg{sjPlot}-functions. See 'Examples'.}

\item{fun}{Plotting function. Refers to the function name of \pkg{sjPlot}-functions.
See 'Details' and 'Examples'.}
}
\value{
NULL.
}
\description{
This function has a pipe-friendly argument-structure, with the
             first argument always being the data, followed by variables that
             should be plotted or printed as table. The function then transforms 
             the input and calls the requested sjp.xy- resp. sjt.xy-function 
             to create a plot or table.
}
\details{
Following \code{fun}-values are currently supported:
         \describe{
            \item{\code{"frq"}}{calls \code{\link{sjp.frq}}. If \code{.data} 
            has more than one variable, a plot for each variable in \code{.data} 
            is plotted.
            }
            \item{\code{"grpfrq"}}{calls \code{\link{sjp.grpfrq}}. The first 
            two variables in \code{.data} are used (and required) to create the plot.
            }
            \item{\code{"xtab"}}{calls \code{\link{sjp.xtab}}. The first 
            two variables in \code{.data} are used (and required) to create the plot.
            }
            \item{\code{"gpt"}}{calls \code{\link{sjp.gpt}}. The first 
            three variables in \code{.data} are used (and required) to create the plot.
            }
            \item{\code{"scatter"}}{calls \code{\link{sjp.scatter}}. The first 
            two variables in \code{.data} are used (and required) to create the plot;
            if \code{.data} also has a third variable, this is used as grouping-
            variable in \code{sjp.scatter}.
            }
         }
}
\note{
The \code{...}-argument is used, first, to specify the variables from \code{.data}
      that should be plotted, and, second, to name further arguments that are
      used in the subsequent plotting functions. Refer to the online-help of
      supported plotting-functions to see valid arguments.
}
\examples{
library(sjmisc)
library(dplyr)
data(efc)

# Frequency plot
sjplot(efc, e42dep, c172code, fun = "frq")

# Grouped frequencies
efc \%>\% sjplot(e42dep, c172code, fun = "grpfrq")

# Grouped frequencies, as box plots
efc \%>\% sjplot(e17age, c172code, fun = "grpfrq", 
               type = "box", geom.colors = "Set1")

# scatter plot, grouped
efc \%>\% 
  select(e42dep, c172code, c161sex) \%>\% 
  sjplot(fun = "scatter")


}

