% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sjs.se}
\alias{sjs.se}
\title{Compute standard error for variables}
\usage{
sjs.se(x)
}
\arguments{
\item{x}{a (numeric) vector / variable.}
}
\value{
The standard error of variable \code{x}.
}
\description{
Compute standard error for variables
}
\examples{
sjs.se(rnorm(n = 100, mean = 3))
}

