% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabFA.R
\name{sjt.fa}
\alias{sjt.fa}
\title{Summary of factor analysis as HTML table}
\usage{
sjt.fa(data, rotation = c("promax", "varimax"), method = c("ml", "minres",
  "wls", "gls", "pa", "minchi", "minrank"), nmbr.fctr = NULL,
  fctr.load.tlrn = 0.1, title = "Factor Analysis", var.labels = NULL,
  wrap.labels = 40, show.cronb = TRUE, show.comm = FALSE,
  altr.row.col = FALSE, digits = 2, CSS = NULL, encoding = NULL,
  file = NULL, use.viewer = TRUE, no.output = FALSE,
  remove.spaces = TRUE)
}
\arguments{
\item{data}{A data frame that should be used to compute a FA, or a \code{\link[psych]{fa}} object.}

\item{rotation}{Rotation of the factor loadings. May be \code{"varimax"} for orthogonal rotation
or \code{"promax"} for oblique transformation (default). Requires the \code{"GPArotation"} package.}

\item{method}{the factoring method to be used. \code{"ml"} will do a maximum likelihood factor analysis (default).
\code{"minres"} will do a minimum residual (OLS),
\code{"wls"} will do a weighted least squares (WLS) solution,
\code{"gls"} does a generalized weighted least squares (GLS),
\code{"pa"} will do the principal factor solution,
\code{"minchi"} will minimize the sample size weighted chi square
when treating pairwise correlations with different number of
subjects per pair. \code{"minrank"} will do a minimum rank factor analysis.}

\item{nmbr.fctr}{Number of factors used for calculating the rotation. By
default, this value is \code{NULL} and the amount of factors is
calculated according to a parallel analysis.}

\item{fctr.load.tlrn}{Specifies the minimum difference a variable needs to have between
factor loadings (components) in order to indicate a clear loading on just one factor and not
diffusing over all factors. For instance, a variable with 0.8, 0.82 and 0.84 factor loading
on 3 possible factors can not be clearly assigned to just one factor and thus would be removed
from the principal component analysis. By default, the minimum difference of loading values
between the highest and 2nd highest factor should be 0.1}

\item{title}{character vector, used as plot title. Depending on plot type and function,
will be set automatically. If \code{title = ""}, no title is printed.
For effect-plots, may also be a character vector of length > 1,
to define titles for each sub-plot or facet.}

\item{var.labels}{Character vector with variable names, which will be used
to label variables in the output.}

\item{wrap.labels}{numeric, determines how many chars of the value, variable or axis
labels are displayed in one line and when a line break is inserted.}

\item{show.cronb}{Logical, if \code{TRUE} (default), the cronbach's alpha value for each factor scale will be calculated,
i.e. all variables with the highest loading for a factor are taken for the
reliability test. The result is an alpha value for each factor dimension.
Only applies when \code{data} is a data frame.}

\item{show.comm}{Logical, if \code{TRUE}, show the communality column in the table.}
}
\value{
Invisibly returns
         \itemize{
           \item the web page style sheet (\code{page.style}),
           \item the web page content (\code{page.content}),
           \item the complete html-output (\code{page.complete}),
           \item the html-table with inline-css for use with knitr (\code{knitr}),
           \item the \code{factor.index}, i.e. the column index of each variable with the highest factor loading for each factor and
           \item the \code{removed.items}, i.e. which variables have been removed because they were outside of the \code{fctr.load.tlrn}'s range.
           }
           for further use.
}
\description{
Performes a factor analysis on a data frame or matrix
               and displays the factors as HTML
               table, or saves them as file. \cr \cr In case a data frame is used as
               parameter, the Cronbach's Alpha value for each factor scale will be calculated,
               i.e. all variables with the highest loading for a factor are taken for the
               reliability test. The result is an alpha value for each factor dimension.
}
\details{
See 'Details' in \code{\link{sjt.frq}}.
}
\note{
This method for factor analysis relies on the functions
      \code{\link[psych]{fa}} and \code{\link[psych]{fa.parallel}} from the psych package.
}
\examples{
\dontrun{
# Data from the EUROFAMCARE sample dataset
library(sjmisc)
library(GPArotation)
data(efc)

# recveive first item of COPE-index scale
start <- which(colnames(efc) == "c82cop1")
# recveive last item of COPE-index scale
end <- which(colnames(efc) == "c90cop9")
# auto-detection of labels
sjt.fa(efc[, start:end])}

}
