% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabGrpmean.R
\name{sjt.grpmean}
\alias{sjt.grpmean}
\alias{sjt.df}
\alias{sjt.mwu}
\title{Deprecated functions}
\usage{
sjt.grpmean(...)

sjt.df(...)

sjt.mwu(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
