% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_model.R
\name{tab_model}
\alias{tab_model}
\title{Print regression models to HTML table}
\usage{
tab_model(..., transform, show.intercept = TRUE, show.est = TRUE,
  show.ci = 0.95, show.hdi50 = TRUE, show.se = NULL, show.std = NULL,
  show.p = TRUE, show.stat = FALSE, show.df = FALSE,
  show.zeroinf = TRUE, show.r2 = TRUE, show.icc = TRUE,
  show.re.var = TRUE, show.fstat = FALSE, show.aic = FALSE,
  show.aicc = FALSE, show.dev = FALSE, show.obs = TRUE, terms = NULL,
  rm.terms = NULL, group.terms = TRUE, order.terms = NULL, title = NULL,
  pred.labels = NULL, dv.labels = NULL, wrap.labels = 25,
  string.pred = "Predictors", string.std = "std. Beta", string.ci = "CI",
  string.se = "std. Error", string.p = "p", string.df = "df",
  string.stat = "Statistic", ci.hyphen = "&nbsp;&ndash;&nbsp;",
  minus.sign = "&#45;", separate.ci.col = TRUE, separate.se.col = TRUE,
  digits = 2, digits.p = 3, emph.p = TRUE, p.val = c("wald", "kr"),
  case = "parsed", auto.label = TRUE, bpe = "median",
  CSS = css_theme("regression"))
}
\arguments{
\item{...}{One or more regression models, including glm's or mixed models.
May also be a \code{list} with fitted models. See 'Examples'.}

\item{transform}{A character vector, naming a function that will be applied
on estimates and confidence intervals. By default, \code{transform} will
automatically use \code{"exp"} as transformation for applicable classes of
regression models (e.g. logistic or poisson regression). Estimates of linear
models remain untransformed. Use \code{NULL} if you want the raw,
non-transformed estimates.}

\item{show.intercept}{Logical, if \code{TRUE}, the intercepts are printed.}

\item{show.est}{Logical, if \code{TRUE}, the estimates are printed.}

\item{show.ci}{Either logical, and if \code{TRUE}, the confidence intervals
is printed to the table; if \code{FALSE}, confidence intervals are
omitted. Or numeric, between 0 and 1, indicating the range of the
confidence intervals.}

\item{show.se}{Logical, if \code{TRUE}, the standard errors are also printed.}

\item{show.std}{Indicates whether standardized beta-coefficients should
also printed, and if yes, which type of standardization is done.
See 'Details'.}

\item{show.p}{Logical, adds asterisks that indicate the significance level of
estimates to the value labels.}

\item{show.stat}{Logical, if \code{TRUE}, the coefficients' test statistic
is also printed.}

\item{show.df}{Logical, if \code{TRUE} and \code{p.val = "kr"}, the p-values
for linear mixed models are based on df with Kenward-Rogers approximation.
These df-values are printed. See \code{\link[sjstats]{p_value}} for details.}

\item{show.zeroinf}{Logical, if \code{TRUE} and model has a zero-inflated
model part, this is also printed to the table.}

\item{show.r2}{Logical, if \code{TRUE}, the r-squared value is also printed.
Depending on the model, these might be pseudo-r-squared values, or Bayesian
r-squared etc. See \code{\link[sjstats]{r2}} for details.}

\item{show.icc}{Logical, if \code{TRUE}, prints the intraclass correlation
coefficient for mixed models. See \code{\link[sjstats]{icc}} for details.}

\item{show.re.var}{Logical, if \code{TRUE}, prints the random effect variances
for mixed models. See \code{\link[sjstats]{re_var}} for details.}

\item{show.dev}{Logical, if \code{TRUE}, shows the deviance of the model.}

\item{terms}{Character vector with names of those terms (variables) that should
be printed in the table. All other terms are removed from the output. If
\code{NULL}, all terms are printed. Note that the term names must match
the names of the model's coefficients. For factors, this means that
the variable name is suffixed with the related factor level, and each
category counts as one term. E.g. \code{rm.terms = "t_name [2,3]"}
would remove the terms \code{"t_name2"} and \code{"t_name3"} (assuming
that the variable \code{t_name} is categorical and has at least
the factor levels \code{2} and \code{3}). Another example for the
\emph{iris}-dataset: \code{terms = "Species"} would not work, instead
you would \code{terms = "Species [versicolor,virginica]"}.}

\item{rm.terms}{Character vector with names that indicate which terms should
be removed from the output Counterpart to \code{terms}. \code{rm.terms =
"t_name"} would remove the term \emph{t_name}. Default is \code{NULL}, i.e.
all terms are used. For factors, levels that should be removed from the plot
need to be explicitely indicated in square brackets, and match the model's
coefficient names, e.g. \code{rm.terms = "t_name [2,3]"} would remove the terms
\code{"t_name2"} and \code{"t_name3"} (assuming that the variable \code{t_name}
was categorical and has at least the factor levels \code{2} and \code{3}).}

\item{group.terms}{Logical, if \code{TRUE} (default), automatically groups table rows with
factor levels of same factor, i.e. predictors of type \code{\link{factor}} will
be grouped, if the factor has more than two levels. Grouping means that a separate headline
row is inserted to the table just before the predictor values.}

\item{order.terms}{Numeric vector, indicating in which order the coefficients
should be plotted. See examples in
\href{../doc/plot_model_estimates.html}{this package-vignette}.}

\item{title}{Character vector, used as plot title. By default,
\code{\link[sjlabelled]{get_dv_labels}} is called to retrieve the label of
the dependent variable, which will be used as title. Use \code{title = ""}
to remove title.}

\item{pred.labels}{Character vector with labels of predictor variables.
If not \code{NULL}, \code{pred.labels} will be used in the first
table column with the predictors' names. By default, if \code{auto.label = TRUE}
and \code{\link[sjlabelled]{get_term_labels}} is called to retrieve the labels
of the coefficients, which will be used as predictor labels.
If \code{pred.labels = ""} or \code{auto.label = FALSE}, the raw
variable names as used in the model formula are used as predictor
labels. If \code{pred.labels} is a named vector, predictor labels (by
default, the names of the model's coefficients) will be matched with the
names of \code{pred.labels}. This ensures that labels always match the
related predictor in the table, no matter in which way the predictors
are sorted. See 'Examples'.}

\item{dv.labels}{Character vector with labels of dependent variables of all
fitted models. See 'Examples'.}

\item{wrap.labels}{Numeric, determines how many chars of the value, variable
or axis labels are displayed in one line and when a line break is inserted.}

\item{string.pred}{Character vector,used as headline for the predictor column.
Default is \code{"Predictors"}.}

\item{string.std}{Character vector, used for the column heading of standardized beta coefficients. Default is \code{"std. Beta"}.}

\item{string.ci}{Character vector, used for the column heading of confidence interval values. Default is \code{"CI"}.}

\item{string.se}{Character vector, used for the column heading of standard error values. Default is \code{"std. Error"}.}

\item{string.p}{Character vector, used for the column heading of p values. Default is \code{"p"}.}

\item{ci.hyphen}{Character vector, indicating the hyphen for confidence interval range.
May be an HTML entity. See 'Examples'.}

\item{minus.sign}{string, indicating the minus sign for negative numbers.
May be an HTML entity. See 'Examples'.}

\item{separate.ci.col}{Logical, if \code{TRUE}, the CI values are shown in
a separate table column.}

\item{separate.se.col}{Logical, if \code{TRUE}, the SE values are shown in
a separate table column.}

\item{digits}{Amount of decimals for estimates}

\item{digits.p}{Amount of decimals for p-values}

\item{emph.p}{Logical, if \code{TRUE}, significant p-values are shown bold faced.}

\item{case}{Desired target case. Labels will automatically converted into the
specified character case. See \code{\link[snakecase]{to_any_case}} for more
details on this argument. By default, if \code{case} is not specified,
it will be set to \code{"parsed"}, unless \code{prefix.labels} is not
\code{"none"}. If \code{prefix.labels} is either \code{"label"} (or
\code{"l"}) or \code{"varname"} (or \code{"v"}) and \code{case} is not
specified, it will be set to \code{NULL} - this is a more convenient
default when prefixing labels.}

\item{auto.label}{Logical, if \code{TRUE} (the default), plot-labels are
based on value and variable labels, if the data is labelled. See
\code{\link[sjlabelled]{get_label}} and
\code{\link[sjlabelled]{get_term_labels}} for details. If \code{FALSE},
original variable names and value labels (factor levels) are used.}

\item{bpe}{For \strong{Stan}-models (fitted with the \pkg{rstanarm}- or
\pkg{brms}-package), the Bayesian point estimate is, by default, the median
of the posterior distribution. Use \code{bpe} to define other functions to
calculate the Bayesion point estimate. \code{bpe} needs to be a character
naming the specific function, which is passed to the \code{fun}-argument in
\code{\link[sjstats]{typical_value}}. So, \code{bpe = "mean"} would
calculate the mean value of the posterior distribution.}
}
\description{
\code{tab_model()} creates HTML tables from regression models.
}
