% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{to_sjPlot}
\alias{to_sjPlot}
\title{Convert a haven-imported data frame to sjPlot format}
\usage{
to_sjPlot(x)
}
\arguments{
\item{x}{a data frame, which contains \code{labelled} class vectors or a single vector
of class \code{labelled}.}
}
\value{
a data frame or single vector (depending on \code{x}) with 'sjPlot' friendly
          vector classes.
}
\description{
This function converts a data frame, which was imported with any of
               \code{haven}'s read functions and contains \code{labelled} class vectors or
               a single vector of type \code{labelled} into an sjPlot friendly data
               frame format, which means that simply all \code{labelled} class
               attributes will be removed, so all vectors / variables will most
               likely become \code{\link{atomic}}.
}
\note{
This function is currently only used to avoid possible compatibility issues
        with \code{labelled} class vectors. Some known issues with \code{labelled}
        class vectors have already been fixed, so it might be that this function
        will become redundant in the future.
}
\seealso{
\itemize{
           \item \href{http://www.strengejacke.de/sjPlot/datainit/}{sjPlot manual: data initialization}
           }
}

